requirements.txt
Pillow
numpy
tensorflow
flask
opencv-python==4.9.0.80